Excepcion de formato datos

+++ FrbaCommerce/Entidades/Ent_Cliente.cs
@@ -10,11 +10,11 @@ namespace FrbaCommerce.Entidades
         public String Nombre { get; set; }
         public String Apellido { get; set; }
         public String Mail { get; set; }
-        public Int64 Dni { get; set; }
+        public Decimal Dni { get; set; }
         public String Fecha_Nac { get; set; }
         public String Dom_Calle { get; set; }
-        public Int64 Nro_Calle { get; set; }
-        public Int64 Piso { get; set; }
+        public Decimal Nro_Calle { get; set; }
+        public Decimal Piso { get; set; }
         public String Dpto { get; set; }
         public String Cod_Postal { get; set; }
         public Int16 Tipo_dni { get; set; }
@@ -22,8 +22,8 @@ namespace FrbaCommerce.Entidades
         public String Localidad { get; set; }
     
     public Ent_Cliente() {}
-    public Ent_Cliente(String pNombre, String pApellido, Int64 pDni, Int16 pTipo_dni, String pMail, String pFecha_Nac,
-        String pDom_Calle, Int64 pNro_Calle, Int64 pPiso, String pDpto, String pCod_Postal, String pTelefono, String pLocalidad)
+    public Ent_Cliente(String pNombre, String pApellido, Decimal pDni, Int16 pTipo_dni, String pMail, String pFecha_Nac,
+        String pDom_Calle, Decimal pNro_Calle, Decimal pPiso, String pDpto, String pCod_Postal, String pTelefono, String pLocalidad)
     {
